phase-5-context-engineering/requirements.txt
# Phase 5: Context Engineering
chromadb>=1.0.0
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<0.4.0
langchain-together>=0.3.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
python-dotenv>=1.0.0
jq>=1.0.0
chainlit>=2.0.0
pypdf>=4.0.0
tiktoken>=0.7.0
